backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
chardet==4.0.0
cloudinary==1.44.1
Django==6.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==2.10
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.25.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==1.26.20
whitenoise==6.9.0
